Fix test file type mismatches for int16_t world dimensions
Update test loop variables from uint32_t to int to match the new
int16_t world width/height types. Also fix bounds checks and function
signatures that were comparing signed/unsigned integers.

Test files updated:
- ClockScenario_test.cpp
- WorldLightCalculator_test.cpp
- Duck_test.cpp, Goose_test.cpp, MultiCellOrganism_test.cpp
- Buoyancy_test.cpp, HorizontalMomentum_test.cpp
- WaterEqualization_test.cpp
- CellTrackerUtil.cpp

Key changes:
- Loop variables: uint32_t -> int for world iteration
- Function params: uint32_t -> int for createTestWorld()
- Bounds checks: Remove unnecessary static_cast<uint32_t>
- CoreControls: Fix WorldResize cast to int16_t

+++ apps/src/core/WorldRigidBodyCalculator.cpp
@@ -94,9 +94,9 @@ std::vector<RigidStructure> WorldRigidBodyCalculator::findAllStructures(const Wo
     const auto& data = world.getData();
     std::unordered_set<Vector2i, Vector2iHash> processed;
 
-    for (uint32_t y = 0; y < data.height; ++y) {
-        for (uint32_t x = 0; x < data.width; ++x) {
-            Vector2i pos{ static_cast<int>(x), static_cast<int>(y) };
+    for (int y = 0; y < data.height; ++y) {
+        for (int x = 0; x < data.width; ++x) {
+            Vector2i pos{ x, y };
             if (processed.count(pos)) {
                 continue;
             }
